Move some fixture useful for functional testing to conftest
